back/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
drf-spectacular==0.27.0
Pillow==10.2.0
python-decouple==3.8
gunicorn==21.2.0
